+++ index.refactoring.html
@@ -8,6 +8,6 @@
 </head>
 <body class="bg-gray-100 p-8">
 <div id="root"></div>
-<script type="module" src="/src/refactoring/main.tsx"></script>
+<script type="module" src="/src/origin/main.tsx"></script>
 </body>
 </html>
