+++ src/origin/components/AdminPage.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
-import { Coupon, Discount, Product } from '../../types';
+import { Coupon, Discount, Product } from '../../types.ts';
+import { useNewCoupon } from '../hooks/useNewCoupon.ts';
 
 interface Props {
   products: Product[];
@@ -9,15 +10,18 @@ interface Props {
   onCouponAdd: (newCoupon: Coupon) => void;
 }
 
-export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, onCouponAdd }: Props) => {
+export const AdminPage = ({
+  products,
+  coupons,
+  onProductUpdate,
+  onProductAdd,
+  onCouponAdd
+}: Props) => {
   const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
   const [editingProduct, setEditingProduct] = useState<Product | null>(null);
-  const [newDiscount, setNewDiscount] = useState<Discount>({ quantity: 0, rate: 0 });
-  const [newCoupon, setNewCoupon] = useState<Coupon>({
-    name: '',
-    code: '',
-    discountType: 'percentage',
-    discountValue: 0
+  const [newDiscount, setNewDiscount] = useState<Discount>({
+    quantity: 0,
+    rate: 0
   });
   const [showNewProductForm, setShowNewProductForm] = useState(false);
   const [newProduct, setNewProduct] = useState<Omit<Product, 'id'>>({
@@ -27,8 +31,10 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
     discounts: []
   });
 
+  const { newCoupon, updateNewCoupon, initNewCoupon } = useNewCoupon();
+
   const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds(prev => {
+    setOpenProductIds((prev) => {
       const newSet = new Set(prev);
       if (newSet.has(productId)) {
         newSet.delete(productId);
@@ -41,7 +47,7 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
 
   // handleEditProduct 함수 수정
   const handleEditProduct = (product: Product) => {
-    setEditingProduct({...product});
+    setEditingProduct({ ...product });
   };
 
   // 새로운 핸들러 함수 추가
@@ -69,7 +75,7 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
   };
 
   const handleStockUpdate = (productId: string, newStock: number) => {
-    const updatedProduct = products.find(p => p.id === productId);
+    const updatedProduct = products.find((p) => p.id === productId);
     if (updatedProduct) {
       const newProduct = { ...updatedProduct, stock: newStock };
       onProductUpdate(newProduct);
@@ -78,7 +84,7 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
   };
 
   const handleAddDiscount = (productId: string) => {
-    const updatedProduct = products.find(p => p.id === productId);
+    const updatedProduct = products.find((p) => p.id === productId);
     if (updatedProduct && editingProduct) {
       const newProduct = {
         ...updatedProduct,
@@ -91,7 +97,7 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
   };
 
   const handleRemoveDiscount = (productId: string, index: number) => {
-    const updatedProduct = products.find(p => p.id === productId);
+    const updatedProduct = products.find((p) => p.id === productId);
     if (updatedProduct) {
       const newProduct = {
         ...updatedProduct,
@@ -104,12 +110,7 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
 
   const handleAddCoupon = () => {
     onCouponAdd(newCoupon);
-    setNewCoupon({
-      name: '',
-      code: '',
-      discountType: 'percentage',
-      discountValue: 0
-    });
+    initNewCoupon();
   };
 
   const handleAddNewProduct = () => {
@@ -140,32 +141,59 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
             <div className="bg-white p-4 rounded shadow mb-4">
               <h3 className="text-xl font-semibold mb-2">새 상품 추가</h3>
               <div className="mb-2">
-                <label htmlFor="productName" className="block text-sm font-medium text-gray-700">상품명</label>
+                <label
+                  htmlFor="productName"
+                  className="block text-sm font-medium text-gray-700"
+                >
+                  상품명
+                </label>
                 <input
                   id="productName"
                   type="text"
                   value={newProduct.name}
-                  onChange={(e) => setNewProduct({ ...newProduct, name: e.target.value })}
+                  onChange={(e) =>
+                    setNewProduct({ ...newProduct, name: e.target.value })
+                  }
                   className="w-full p-2 border rounded"
                 />
               </div>
               <div className="mb-2">
-                <label htmlFor="productPrice" className="block text-sm font-medium text-gray-700">가격</label>
+                <label
+                  htmlFor="productPrice"
+                  className="block text-sm font-medium text-gray-700"
+                >
+                  가격
+                </label>
                 <input
                   id="productPrice"
                   type="number"
                   value={newProduct.price}
-                  onChange={(e) => setNewProduct({ ...newProduct, price: parseInt(e.target.value) })}
+                  onChange={(e) =>
+                    setNewProduct({
+                      ...newProduct,
+                      price: parseInt(e.target.value)
+                    })
+                  }
                   className="w-full p-2 border rounded"
                 />
               </div>
               <div className="mb-2">
-                <label htmlFor="productStock" className="block text-sm font-medium text-gray-700">재고</label>
+                <label
+                  htmlFor="productStock"
+                  className="block text-sm font-medium text-gray-700"
+                >
+                  재고
+                </label>
                 <input
                   id="productStock"
                   type="number"
                   value={newProduct.stock}
-                  onChange={(e) => setNewProduct({ ...newProduct, stock: parseInt(e.target.value) })}
+                  onChange={(e) =>
+                    setNewProduct({
+                      ...newProduct,
+                      stock: parseInt(e.target.value)
+                    })
+                  }
                   className="w-full p-2 border rounded"
                 />
               </div>
@@ -179,7 +207,11 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
           )}
           <div className="space-y-2">
             {products.map((product, index) => (
-              <div key={product.id} data-testid={`product-${index + 1}`} className="bg-white p-4 rounded shadow">
+              <div
+                key={product.id}
+                data-testid={`product-${index + 1}`}
+                className="bg-white p-4 rounded shadow"
+              >
                 <button
                   data-testid="toggle-button"
                   onClick={() => toggleProductAccordion(product.id)}
@@ -196,7 +228,12 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
                           <input
                             type="text"
                             value={editingProduct.name}
-                            onChange={(e) => handleProductNameUpdate(product.id, e.target.value)}
+                            onChange={(e) =>
+                              handleProductNameUpdate(
+                                product.id,
+                                e.target.value
+                              )
+                            }
                             className="w-full p-2 border rounded"
                           />
                         </div>
@@ -205,7 +242,12 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
                           <input
                             type="number"
                             value={editingProduct.price}
-                            onChange={(e) => handlePriceUpdate(product.id, parseInt(e.target.value))}
+                            onChange={(e) =>
+                              handlePriceUpdate(
+                                product.id,
+                                parseInt(e.target.value)
+                              )
+                            }
                             className="w-full p-2 border rounded"
                           />
                         </div>
@@ -214,18 +256,33 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
                           <input
                             type="number"
                             value={editingProduct.stock}
-                            onChange={(e) => handleStockUpdate(product.id, parseInt(e.target.value))}
+                            onChange={(e) =>
+                              handleStockUpdate(
+                                product.id,
+                                parseInt(e.target.value)
+                              )
+                            }
                             className="w-full p-2 border rounded"
                           />
                         </div>
                         {/* 할인 정보 수정 부분 */}
                         <div>
-                          <h4 className="text-lg font-semibold mb-2">할인 정보</h4>
+                          <h4 className="text-lg font-semibold mb-2">
+                            할인 정보
+                          </h4>
                           {editingProduct.discounts.map((discount, index) => (
-                            <div key={index} className="flex justify-between items-center mb-2">
-                              <span>{discount.quantity}개 이상 구매 시 {discount.rate * 100}% 할인</span>
+                            <div
+                              key={index}
+                              className="flex justify-between items-center mb-2"
+                            >
+                              <span>
+                                {discount.quantity}개 이상 구매 시{' '}
+                                {discount.rate * 100}% 할인
+                              </span>
                               <button
-                                onClick={() => handleRemoveDiscount(product.id, index)}
+                                onClick={() =>
+                                  handleRemoveDiscount(product.id, index)
+                                }
                                 className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
                               >
                                 삭제
@@ -237,14 +294,24 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
                               type="number"
                               placeholder="수량"
                               value={newDiscount.quantity}
-                              onChange={(e) => setNewDiscount({ ...newDiscount, quantity: parseInt(e.target.value) })}
+                              onChange={(e) =>
+                                setNewDiscount({
+                                  ...newDiscount,
+                                  quantity: parseInt(e.target.value)
+                                })
+                              }
                               className="w-1/3 p-2 border rounded"
                             />
                             <input
                               type="number"
                               placeholder="할인율 (%)"
                               value={newDiscount.rate * 100}
-                              onChange={(e) => setNewDiscount({ ...newDiscount, rate: parseInt(e.target.value) / 100 })}
+                              onChange={(e) =>
+                                setNewDiscount({
+                                  ...newDiscount,
+                                  rate: parseInt(e.target.value) / 100
+                                })
+                              }
                               className="w-1/3 p-2 border rounded"
                             />
                             <button
@@ -266,7 +333,10 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
                       <div>
                         {product.discounts.map((discount, index) => (
                           <div key={index} className="mb-2">
-                            <span>{discount.quantity}개 이상 구매 시 {discount.rate * 100}% 할인</span>
+                            <span>
+                              {discount.quantity}개 이상 구매 시{' '}
+                              {discount.rate * 100}% 할인
+                            </span>
                           </div>
                         ))}
                         <button
@@ -292,20 +362,29 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
                 type="text"
                 placeholder="쿠폰 이름"
                 value={newCoupon.name}
-                onChange={(e) => setNewCoupon({ ...newCoupon, name: e.target.value })}
+                onChange={(e) =>
+                  updateNewCoupon({ ...newCoupon, name: e.target.value })
+                }
                 className="w-full p-2 border rounded"
               />
               <input
                 type="text"
                 placeholder="쿠폰 코드"
                 value={newCoupon.code}
-                onChange={(e) => setNewCoupon({ ...newCoupon, code: e.target.value })}
+                onChange={(e) =>
+                  updateNewCoupon({ ...newCoupon, code: e.target.value })
+                }
                 className="w-full p-2 border rounded"
               />
               <div className="flex gap-2">
                 <select
                   value={newCoupon.discountType}
-                  onChange={(e) => setNewCoupon({ ...newCoupon, discountType: e.target.value as 'amount' | 'percentage' })}
+                  onChange={(e) =>
+                    updateNewCoupon({
+                      ...newCoupon,
+                      discountType: e.target.value as 'amount' | 'percentage'
+                    })
+                  }
                   className="w-full p-2 border rounded"
                 >
                   <option value="amount">금액(원)</option>
@@ -315,7 +394,12 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
                   type="number"
                   placeholder="할인 값"
                   value={newCoupon.discountValue}
-                  onChange={(e) => setNewCoupon({ ...newCoupon, discountValue: parseInt(e.target.value) })}
+                  onChange={(e) =>
+                    updateNewCoupon({
+                      ...newCoupon,
+                      discountValue: parseInt(e.target.value)
+                    })
+                  }
                   className="w-full p-2 border rounded"
                 />
               </div>
@@ -330,9 +414,16 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
               <h3 className="text-lg font-semibold mb-2">현재 쿠폰 목록</h3>
               <div className="space-y-2">
                 {coupons.map((coupon, index) => (
-                  <div key={index} data-testid={`coupon-${index + 1}`} className="bg-gray-100 p-2 rounded">
+                  <div
+                    key={index}
+                    data-testid={`coupon-${index + 1}`}
+                    className="bg-gray-100 p-2 rounded"
+                  >
                     {coupon.name} ({coupon.code}):
-                    {coupon.discountType === 'amount' ? `${coupon.discountValue}원` : `${coupon.discountValue}%`} 할인
+                    {coupon.discountType === 'amount'
+                      ? `${coupon.discountValue}원`
+                      : `${coupon.discountValue}%`}{' '}
+                    할인
                   </div>
                 ))}
               </div>
