+++ src/origin/components/CartPage.tsx
@@ -1,5 +1,5 @@
-import { useState } from 'react';
 import { CartItem, Coupon, Product } from '../../types.ts';
+import { useCart } from "../hooks";
 
 interface Props {
   products: Product[];
@@ -7,79 +7,15 @@ interface Props {
 }
 
 export const CartPage = ({ products, coupons }: Props) => {
-  const [cart, setCart] = useState<CartItem[]>([]);
-  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
-
-
-  const addToCart = (product: Product) => {
-    const remainingStock = getRemainingStock(product);
-    if (remainingStock <= 0) return;
-
-    setCart(prevCart => {
-      const existingItem = prevCart.find(item => item.product.id === product.id);
-      if (existingItem) {
-        return prevCart.map(item =>
-          item.product.id === product.id
-            ? { ...item, quantity: Math.min(item.quantity + 1, product.stock) }
-            : item
-        );
-      }
-      return [...prevCart, { product, quantity: 1 }];
-    });
-  };
-
-  const removeFromCart = (productId: string) => {
-    setCart(prevCart => prevCart.filter(item => item.product.id !== productId));
-  };
-
-  const updateQuantity = (productId: string, newQuantity: number) => {
-    setCart(prevCart =>
-      prevCart.map(item => {
-        if (item.product.id === productId) {
-          const maxQuantity = item.product.stock;
-          const updatedQuantity = Math.max(0, Math.min(newQuantity, maxQuantity));
-          return updatedQuantity > 0 ? { ...item, quantity: updatedQuantity } : null;
-        }
-        return item;
-      }).filter((item): item is CartItem => item !== null)
-    );
-  };
-
-  const calculateTotal = () => {
-    let totalBeforeDiscount = 0;
-    let totalAfterDiscount = 0;
-
-    cart.forEach(item => {
-      const { price } = item.product;
-      const { quantity } = item;
-      totalBeforeDiscount += price * quantity;
-
-      const discount = item.product.discounts.reduce((maxDiscount, d) => {
-        return quantity >= d.quantity && d.rate > maxDiscount ? d.rate : maxDiscount;
-      }, 0);
-
-      totalAfterDiscount += price * quantity * (1 - discount);
-    });
-
-    let totalDiscount = totalBeforeDiscount - totalAfterDiscount;
-
-    // 쿠폰 적용
-    if (selectedCoupon) {
-      if (selectedCoupon.discountType === 'amount') {
-        totalAfterDiscount = Math.max(0, totalAfterDiscount - selectedCoupon.discountValue);
-      } else {
-        totalAfterDiscount *= (1 - selectedCoupon.discountValue / 100);
-      }
-      totalDiscount = totalBeforeDiscount - totalAfterDiscount;
-    }
-
-    return {
-      totalBeforeDiscount: Math.round(totalBeforeDiscount),
-      totalAfterDiscount: Math.round(totalAfterDiscount),
-      totalDiscount: Math.round(totalDiscount)
-    };
-  };
-
+  const {
+    cart,
+    addToCart,
+    removeFromCart,
+    updateQuantity,
+    applyCoupon,
+    calculateTotal,
+    selectedCoupon
+  } = useCart();
 
   const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
     return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
@@ -102,10 +38,6 @@ export const CartPage = ({ products, coupons }: Props) => {
       }
     }
     return appliedDiscount;
-  };
-
-  const applyCoupon = (coupon: Coupon) => {
-    setSelectedCoupon(coupon);
   };
 
   return (
