+++ src/origin/hooks/useCart.ts
@@ -1,7 +1,10 @@
 // useCart.ts
 import { useState } from 'react';
-import { CartItem, Coupon, Product } from '../../types';
-import { calculateCartTotal, updateCartItemQuantity } from './utils/cartUtils';
+import { CartItem, Coupon, Product } from '../../types.ts';
+import {
+  calculateCartTotal,
+  updateCartItemQuantity
+} from './utils/cartUtils.ts';
 
 export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
