+++ src/origin/stores/useProductStore.ts
@@ -1,5 +1,5 @@
 import { create } from 'zustand';
-import { Product } from '../../types';
+import { Product } from '../../types.ts';
 
 interface Store {
   products: Product[];
